backend/requirements.txt
fastapi
uvicorn
pymongo
motor
boto3
python-multipart
python-dotenv
pydantic
requests
pytest
pandas
loguru